Update types, separate components Add update utils to user
