+++ PasswordMenager/frontend/src/components/AppBar/AccountContainer/AccountInfo/UserView/index.tsx
@@ -1,137 +1,18 @@
 import { useState } from "react";
 import { ModalOpenUtils } from "../../../../../utils/moda.open.utils";
-import Button from "../../../../Button";
-import FormElement from "../../../../FormElement";
 import Modal from "../../../../Modal";
 import { Translation } from "../../../../Translation";
 import { EditIcon } from "../../../../icons/EditIcon";
-import { MoreOptionIcon } from "../../../../icons/MoreOption";
 import {
-  FormElements,
-  IconHoover,
-  PasswordText,
-  UserEditModalButtons,
-  UserEditModalContainer,
-  UserEditModalView,
-  UserEditTitle,
   UserIcons,
   UserInfo,
   UserInfoFlexContainer,
   UserInforContainer,
   UserTitleText,
 } from "./component.styled";
+import { UserEditModalComponent } from "./userEditModal";
+import { User } from "../../../../../utils/user.utils";
 
-const UserEditModalComponent = ({ user, onClose }: any) => {
-  const [login, setLogin] = useState<string>(user.login);
-  const [email, setEmail] = useState<string>(user.email);
-  const [password, setPassword] = useState<string | undefined>(undefined);
-  const [isAdvanceOptionVisible, setAdvanceOption] = useState<boolean>(false);
-
-  return (
-    <UserEditModalView>
-      <UserEditTitle>
-        <span>
-          {Translation("editusermodal.importpassword.infomodal.title")}
-        </span>
-        <IconHoover>
-          <MoreOptionIcon
-            click={() =>
-              !isAdvanceOptionVisible
-                ? setAdvanceOption(true)
-                : setAdvanceOption(false)
-            }
-          />
-          <span test-tooltip="test">
-            {Translation("editusermodal.importpassword.infomodal.moreOption")}
-          </span>
-        </IconHoover>
-      </UserEditTitle>
-      <UserEditModalContainer>
-        <FormElements>
-          <FormElement
-            label={"userinformation.edit.login.label"}
-            inputplaceholder="userinformation.edit.login.label"
-            inputChange={(e) => {
-              setLogin(e.target.value);
-            }}
-            inputtype="txt"
-            fontSize="18px"
-            width="80%"
-            inputFontSize="16px"
-            value={login}
-          />
-          <FormElement
-            label={"userinformation.edit.email.label"}
-            inputplaceholder="userinformation.edit.email.label"
-            inputChange={(e) => {
-              setEmail(e.target.value);
-            }}
-            inputtype="txt"
-            fontSize="18px"
-            width="80%"
-            inputFontSize="16px"
-            value={email}
-          />
-        </FormElements>
-        {isAdvanceOptionVisible ? (
-          <FormElements>
-            <FormElement
-              withTooltip={true}
-              tooltipText="editusermodal.importpassword.infomodal"
-              label={"userinformation.edit.importPassword.label"}
-              inputplaceholder="userinformation.edit.importPassword.label"
-              inputChange={(e) => {}}
-              inputtype="txt"
-              fontSize="18px"
-              width="80%"
-              inputFontSize="16px"
-              value={""}
-            />
-          </FormElements>
-        ) : null}
-        <PasswordText>
-          {Translation(
-            "editusermodal.importpassword.infomodal.editPasswordTest"
-          )}
-        </PasswordText>
-        <FormElements>
-          <FormElement
-            label={"userinformation.edit.password.label"}
-            inputplaceholder="userinformation.edit.password.placeholder"
-            inputChange={(e) => {
-              setPassword(e.target.value);
-            }}
-            fontSize="16px"
-            width="80%"
-            inputFontSize="16px"
-            inputtype="password"
-            value={password}
-          />
-          <FormElement
-            label={"userinformation.edit.confirmpassword.label"}
-            inputplaceholder="userinformation.edit.confirmpassword.label"
-            inputChange={() => {}}
-            fontSize="16px"
-            width="80%"
-            inputFontSize="16px"
-            inputtype="password"
-            value={""}
-          />
-        </FormElements>
-        <UserEditModalButtons>
-          <Button onClick={onClose}>
-            {Translation(
-              "editusermodal.importpassword.infomodal.button.cancel"
-            )}
-          </Button>
-          <Button>
-            {Translation("editusermodal.importpassword.infomodal.button.save")}
-          </Button>
-        </UserEditModalButtons>
-      </UserEditModalContainer>
-    </UserEditModalView>
-  );
-};
 export const UserView = ({
   user,
 }: {
@@ -147,13 +28,24 @@ export const UserView = ({
     setUserModalVisible(false);
     ModalOpenUtils.getInstance().CloseModal = false;
   };
+
+  const saveUserHandle = (user: UserUpdate) => {
+    const isLoginEmpty = 'login' in user && user.login === '';
+    const isEmailEmpty = 'email' in user && user.email === '';
+    if(isLoginEmpty || isEmailEmpty) return;
+    User.getInstance().updateUser(user);
+  };
   return (
     <>
       <Modal
         visible={userEditModalVisible}
         onClose={() => closeCheck()}
         component={
-          <UserEditModalComponent user={user} onClose={() => closeCheck()} />
+          <UserEditModalComponent
+            user={user}
+            onClose={() => closeCheck()}
+            changeHandle={saveUserHandle}
+          />
         }
       ></Modal>
       <UserInfoFlexContainer>
