+++ PasswordMenager/frontend/src/utils/user.utils.ts
@@ -25,4 +25,9 @@ export class User {
     const token = this.sessionStorage.getAccessToken();
     return this.user.getUserInfo(token).then((data) => data.json());
   }
+
+  updateUser(user: UserUpdate): Promise<unknown> {
+    const token = this.sessionStorage.getAccessToken();
+    return this.user.updateUserInfo(token, user);
+  }
 }
